Refactored the process_one_image() function to align with the Inputs that Zach was already sending to us. Take Data and Label nhdr file (paths to file) as input, instead of taking the input directory as input and searching.
Added a new function find_nhdr_file() to search for nhdr file (locally). This is useful because if we are given a COLOR file to process, then we need to find the DWI nhdr file from which to pull relevant transform/resolution metadata from.
This function is written to search the s3 bucket also (when not given a local search dir or when not found locally), but this thought is incomplete. I need to add a function to load an nhdr in (using pynrrd) from s3. This will require using subprocess.run() with rclone "cat" function, and pipe the stdout to a string to give to pynrrd